pyproject.toml
[tool.poetry]
name = "billnet-service"
version = "0.1.0"
description = "A comprehensive event management system API"
authors = ["Syed Tazyeen <themesyed@gmail.com>"]
readme = "README.md"
packages = [{include = "apps"}, {include = "config"}, {include = "core"}]

[tool.poetry.dependencies]
python = "^3.11.0,<3.13"
Django = "^4.2.7"
djangorestframework = "^3.14.0"
drf-spectacular = "^0.26.5"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
django-cors-headers = "^4.3.1"
python-decouple = "^3.8"
PyJWT = "^2.8.0"
psycopg2-binary = "^2.9.9"
djangorestframework-simplejwt = "^5.3.0"
Pillow = "^10.0.0"
requests = "^2.31.0"
crewai = "^0.28.0"
google-generativeai = "^0.3.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
pytest-django = "^4.5.0"
black = "^23.0.0"
flake8 = "^6.0.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 100

[tool.poetry.scripts]
start = "run:main"
makemigrations = "manage:makemigrations"
migrate = "manage:migrate"
